refactored pipeline tests again & fixed fixture scoping
